DECO-137 Move useful code to set system time from SNTP request to SNTP library
